+++ backend/app/Hooks/LicenseSystem.php
@@ -76,8 +76,14 @@ class LicenseSystem
         }
 
         try {
-            $client = new Client();
-            $response = $client->get(self::API_BASE_URL . '/license/' . $licenseKey . '/info');
+            $client = new Client([
+                'force_ip_resolve' => 'v4'
+            ]);
+            $response = $client->get(self::API_BASE_URL . '/license/' . $licenseKey . '/info', [
+                'headers' => [
+                    'MythicalDash-Install' => 'true'
+                ]
+            ]);
 
             $data = json_decode($response->getBody()->getContents(), true);